Pug-Backend/requirements.txt
pymysql
mysqlclient
djangorestframework
django-cors-headers
djoser